refactor(filter-select): rename docs folder to _docs
